MANIFEST: Add BroadcastRecv for wifi events
This patch adds the BroadcastReceiver for wifi connection
events in the manifest file. Every connection or disconnection
will call the WifiReceiver, which will do checkIn or checkOut,
if necessary. The BroadcastReceiver is placed in the manifest
instead of in an Activity because we want this receiver to not
get killed with the app, when the user closes it.